refactor(eval): expose lightweight baseline contracts
Move the dependency-free localization task, location, result, and scoring contracts to codenib.eval.baseline while preserving the agent_runner import path as a compatibility re-export. Include instance and repository identity in agent call context.

Verified with: pytest -q test/eval/test_baseline.py
