+++ codenib/eval/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2025-2026 CodeNib Contributors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Evaluation harnesses: agent runners, retrieval metrics, artifact bundles."""
